refacto(web-sockets): Move handlers in their own file
