+++ packages/server/src/modules/websocket/initWebSocket.ts
@@ -1,8 +1,8 @@
 import { Express } from "express";
 import { createServer } from "http";
 import { Server } from "socket.io";
-import { z } from "zod";
-import { services as gameServices } from "../games/services";
+import { handleJoinGame } from "./eventHandlers/joinGameHandler";
+import { handleUpdateGame } from "./eventHandlers/updateGameHandler";
 
 export { initWebSocket };
 
@@ -11,28 +11,8 @@ function initWebSocket({ app }: { app: Express }) {
   const io = new Server(httpServer, {});
 
   io.on("connection", (socket) => {
-    socket.on("joinGame", (rawGameId: unknown) => {
-      const gameId = z.number().parse(rawGameId);
-      socket.join(`${gameId}`);
-      gameServices
-        .initState({ gameId })
-        .then((state) => socket.emit("resetGameState", state));
-    });
-    socket.on("updateGame", (args: unknown) => {
-      const schema = z.object({
-        gameId: z.number(),
-        update: z.object({
-          step: z.number().optional(),
-          isStepActive: z.boolean().optional(),
-        }),
-      });
-      const { gameId, update } = schema.parse(args);
-      gameServices.update(gameId, update).then(() => {
-        socket.broadcast
-          .to(`${gameId}`)
-          .emit("gameUpdated", { gameId, update });
-      });
-    });
+    handleJoinGame(socket);
+    handleUpdateGame(socket);
   });
   return { httpServer };
 }
